Add provisioned?/1 and hide configure
If a NervesKey starts returning errors, it seems to be since I forgot to
provision it. Add a helper method to make this easier.

This also makes the `configure` method private since it shouldn't be
called externally and in retrospect, it's pretty confusingly named
unless you've read the ATECC508A data sheet.