Support >16 byte serial numbers
This makes use of the flags field to indicate that the rest of the OTP
area should be used for the serial number. This lets you have up to 48
byte serial numbers which certainly seems like way more than anyone
would ever need.

This change also beefs up the OTP section tests to ensure that the code
is backwards compatible since it's assumed that most users will continue
to have <= 16 byte serial numbers.