Add volatile key configuration as an option.
This configuration locks down the private key and an AES encryption key
using the ATECC608 persistent latch and volatile key mechanisms. The
device is provisioned with an activation key that needs to be kept for
unlocking the device.

This allows implementing secret storage as long as you can physically
protect the device and manage power to it using a tamper switch or similar.
The authorization is kept in volatile memory that will survive sleep but
not losing power.

Also adds livebooks for provisioning and testing volatile keys.

Drop Elixir 1.14 support.

Update dependency on atecc508a to 1.4